+++ src/main/java/com/kh/clendy/admin/model/vo/Seller.java
@@ -8,7 +8,7 @@ public class Seller {
 	private String seller_name;
 	private String rep_name;
 	private int rep_num;
-	private int b_license;
+	private long b_license;
 	private String o_b_license;
 	private int phone;
 	private String email;
@@ -22,7 +22,7 @@ public class Seller {
 	
 	public Seller() {}
 
-	public Seller(int seller_code, int a_no, String seller_name, String rep_name, int rep_num, int b_license,
+	public Seller(int seller_code, int a_no, String seller_name, String rep_name, int rep_num, long b_license,
 			String o_b_license, int phone, String email, String address, String account_bank, int account_num,
 			Date apply_date, int user_no, int sales, int commission) {
 		super();
@@ -84,11 +84,11 @@ public class Seller {
 		this.rep_num = rep_num;
 	}
 
-	public int getB_license() {
+	public long getB_license() {
 		return b_license;
 	}
 
-	public void setB_license(int b_license) {
+	public void setB_license(long b_license) {
 		this.b_license = b_license;
 	}
